may het pin

+++ detectnoise.py
@@ -68,10 +68,10 @@ def algorithm(windowsize,img):
         for j in range(w-windowsize+1):
             if caculateT1(windowsize,(i,j),img[:,:,0],img[i,j,0]) and caculateT1(windowsize,(i,j),img[:,:,1],img[i,j,1]) and caculateT1(windowsize,(i,j),img[:,:,2],img[i,j,2]):
                 #noise_index = np.append(noise_index,[[i + (windowsize/2),j + (windowsize/2)]])
-                rows = np.append(rows,[i + (windowsize/2)])
-                cols = np.append(cols,[j + (windowsize/2)])
+                rows = np.append(rows,[i + int(windowsize/2)])
+                cols = np.append(cols,[j + int(windowsize/2)])
                 sumnoise+=1
-    return (rows,cols)     
+    return int(rows),int(cols)
